+++ src/ProgressIndicator/CircularProgress/CircularProgress.tsx
@@ -44,7 +44,7 @@ export enum CircularProgressTestIds {
  */
 export type CircularProgressProps = MuiCircularProgressProps & {
   showprogress?: boolean | 0 | 1,
-  withbackdrop?: boolean,
+  withbackdrop?: boolean | 0 | 1,
 }
 
 export const getMuiCircularProgressThemeOverrides = (): Components<Omit<Theme, 'components'>> => {
@@ -54,11 +54,11 @@ export const getMuiCircularProgressThemeOverrides = (): Components<Omit<Theme, '
         root: ({ ownerState, theme }) => {
           return ({
             ...theme.typography.body2,
-            color: ownerState.withbackdrop ? theme.palette.info.inverse : theme.palette.primary.main,
+            color: (ownerState.withbackdrop === 1 || ownerState.withbackdrop === true) ? theme.palette.info.inverse : theme.palette.primary.main,
             position: 'absolute',
             left: 0,
             ...(ownerState.variant === CircularProgressVariants.DETERMINATE && ownerState.id === CircularProgressTestIds.PROGRESS_TRAIL) && {
-              color: ownerState.withbackdrop ? theme.palette.action.hoverInverse : theme.palette.action.hover,
+              color: (ownerState.withbackdrop === 1 || ownerState.withbackdrop === true) ? theme.palette.action.hoverInverse : theme.palette.action.hover,
             },
             ...ownerState.variant === CircularProgressVariants.DETERMINATE && {
               '+ .MuiBox-root': { // styles the label element right next to a determinate progress circle